Clean up options processing and validation
This removes a chunk of boilerplate options handling code in main.go.

By making better use of existing validators, we are also able to simplify
the parameter validation. We can use the `url` validator for both
`proxy-prometheus-base-url` and `proxy-cors-allowed-origins`, and we can
combine it with other validators to achieve the restrictions we need.